requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
rembg>=2.0.50
torch>=1.9.0
transformers>=4.15.0
mediapipe>=0.8.9
requests>=2.25.1
torchvision>=0.10.0
clipseg
